Add support for methods from `react-native-edge-to-edge`
